Explore, Explain and Examine Predictive Models

This function finds Variable Attributions via Sequential Variable Conditioning. It calls either local_attributions for additive attributions or local_interactions for attributions with interactions.

Details

Instance-level explainers

Dataset-level explainers

Further Reading

Explanatory Model Analysis book

References

https://modeloriented.github.io/iBreakDown/reference/break_down.html

https://modeloriented.github.io/ingredients/reference/ceteris_paribus.html

Methods

Public methods

Explanations$new()

Explanations$plot_break_down()

Explanations$plot_ceteris_paribus()

Explanations$clone()

Method new()

Construct an Explanations object

Usage

Explanations$new(...)

Arguments

...

(ModelComposition) One or more objects created by ModelComposition.

Method plot_break_down()

Usage

Explanations$plot_break_down(new_observation, ...)

Arguments

new_observation

(data.frame) A new observation with columns that correspond to variables used in the model.

...

parameters passed to break_down.

Method plot_ceteris_paribus()

Usage

Explanations$plot_ceteris_paribus(new_observation, variables, ...)

Arguments

new_observation

(data.frame) A new observation with columns that correspond to variables used in the model.

variables

(character) A character vector with the variables names to analyse.

...

parameters passed to ceteris_paribus.

Method clone()

The objects of this class are cloneable with this method.

Usage

Explanations$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.